+++ src/command_modules/azure-cli-lab/azure/cli/command_modules/lab/help.py
@@ -147,3 +147,18 @@ helps['lab formula'] = """
             type: group
             short-summary: Commands to manage formulas of a DevTest Lab.
             """
+helps['lab secret'] = """
+            type: group
+            short-summary: Commands to manage secrets of a DevTest Lab.
+            """
+helps['lab secret set'] = """
+            type: command
+            short-summary: Sets a secret in the DevTest Lab.
+            parameters:
+                - name: --lab-name
+                  short-summary: Name of the Lab
+                - name: --name -n
+                  short-summary: Name of the secret
+                - name: --value
+                  short-summary: Value of the secret
+            """