+++ src/command_modules/azure-cli-lab/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
++++++++++++++++++++++
+
+* Add commands to manage secrets within a Lab.
+
 0.0.1 (2017-04-03)
 +++++++++++++++++++++